Consolidating the move function
Also removed the position setting in SurveyEdit as it is now taken care
of by move. This brings it inline with form editing.
